fix: add ::numeric cast to all ROUND(AVG()) SQL queries
PostgreSQL does not have ROUND(double precision, integer) — the numeric
variant is required. Fix all occurrences across dashboard, dev_auth,
developers, devops_agent, reports, and requests routers.

Also fix uuid = text type mismatch in team-efficiency report JOIN.

+++ services/admin/app/routers/dashboard.py
@@ -12,7 +12,7 @@ _STATS_QUERY = text("""
         COUNT(cr.id) AS request_count,
         SUM(cr.tokens_input + cr.tokens_output) AS total_tokens,
         ROUND(SUM(cr.cost_usd)::numeric, 4) AS total_cost_usd,
-        ROUND(AVG(CASE WHEN cr.cache_hit THEN 1.0 ELSE 0.0 END) * 100, 1) AS cache_hit_pct
+        ROUND((AVG(CASE WHEN cr.cache_hit THEN 1.0 ELSE 0.0 END) * 100)::numeric, 1) AS cache_hit_pct
     FROM teams t
     LEFT JOIN cost_records cr ON cr.team_id = t.id
     GROUP BY t.id, t.name
