+++ services/admin/app/routers/developers.py
@@ -109,8 +109,8 @@ async def get_developer_stats(
             COALESCE(ROUND(SUM(cr.cost_usd)::numeric, 6), 0)             AS cost_usd,
             COALESCE(SUM(cr.tool_invocation_count), 0)                   AS tool_invocations,
             COALESCE(SUM(cr.retry_count), 0)                             AS retry_count,
-            ROUND(AVG(CASE WHEN cr.cache_hit THEN 1.0 ELSE 0.0 END)*100, 1) AS cache_hit_pct,
-            ROUND(AVG(cr.latency_ms), 0)                                 AS avg_latency_ms,
+            ROUND((AVG(CASE WHEN cr.cache_hit THEN 1.0 ELSE 0.0 END)*100)::numeric, 1) AS cache_hit_pct,
+            ROUND(AVG(cr.latency_ms)::numeric, 0)                                 AS avg_latency_ms,
             COUNT(CASE WHEN cr.request_error_type IS NOT NULL THEN 1 END) AS error_count
         FROM cost_records cr
         WHERE cr.developer_id = :dev_id {since}
@@ -282,7 +282,7 @@ async def at_risk_developers(
         ),
         session_quality AS (
             SELECT s.developer_id,
-                   ROUND(AVG(s.quality_score), 2) AS avg_quality,
+                   ROUND(AVG(s.quality_score)::numeric, 2) AS avg_quality,
                    COUNT(CASE WHEN s.produced_commit THEN 1 END) AS sessions_with_commit,
                    COUNT(s.session_trace_id) AS session_count
             FROM sessions s
